requirements/dev.txt
# Specifies only dev-specific requirements
# But imports the common ones too
-r common.txt

django-debug-toolbar==4.2.0